requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
selenium==4.15.2
python-dotenv==1.0.0
google-generativeai==0.3.0
